Deploy scripts (only tested on local net)
It remains to be seen if `make` will make the cut. The invocations are a
little awkward, so it currently only serves to enshrine them. Call with:

* `make deploy-omtf`
* `make deploy-omt owner=0xa0ee7a142d267c1f36714e4a8f75612f20a79720 name='Deciduous\ Tree\ DAO\ Membership' symbol=DTDM base-uri="ipfs://deadb33f"`

the space escaping for the `name` param is unfortunately necessary for
make and may be the reason for not continuing to use it.

Currently on local net, these contract deploys are not verifying.
Unclear whether this is just because it's on local or not, so this will
need to be validated by deploying to a testnet. Currently the make
commands are hardcoded to use the local anvil network.