Upgrade to 0.8.17 (and delete old versions)
In order to _not_ delete the old versions, we would have had to upgrade
their compiler version so as not to get errors... which sort of defeats
the point. It was a sort of silly and high maintenance idea in the first
place considering we already have source control.

Additionally, this does have a small code change to specify some
interfaces that overrides apply to, since the new compiler version
called them out as errors.