Extend Governor to support proposal-based params
This will let us pass additional data so that we can inform voting and
counting strategies on a per-proposal basis as opposed to only being
able to set this on the Governor itself.

GovernorWithProposalParams supplies default proposal params that make
the voting strategy use an IVotes-compatible NFT. In our case, this will
use the OrigamiMembershipToken which will effectively make the default
strategy 1 person, 1 vote, since our tokens are soulbound singletons.