Add but don't yet use GovernorSettingsFacet
IGovernorUpgradeable is still being implemented by OrigamiGovernor, which
has conflicting types with this implementation. Re-typing it to match
IGovernorUpgradeable also doesn't fix the problem since it would require
pulling in and overriding GovernorSettingsUpradeable as well and would
start coupling this facet to those implementations... better to leave
the integration out for the moment.