Jib script to grant application permissions
For each address, grant it revoker and minter roles. The role encodings
are static for that contract, so they're just set to variables here.

A nice optimization would be to combine this all into a single
transaction. For sake of expedience, this just uses a loop with two
calls per address... since we're using wallet pool sizes of 10, this
needs 20 transactions and takes a bit less than 3 minutes on Polygon.