Extract base governance token contract
The main thrust of this was to pull out the primary functionality of our
ERC20 token into a separate contract so it could be deployed
independently if desirable.

The storage changes are fine since _l1Token should never have been in
there and since the transfer lock is now using diamond storage.