Deploy multiple DAOs with `local-deploy` (#34)
* Re-use implemetnation contracts between local deploys
* Check that a contract isn't already deployed at that address
* Minor refactors for clarity
* Make local-deploy per-run configurable
* emit an artifacts file with the addresses of contracts created that run
* Use GovernorSettings struct for governor init

This puts a very permanent end to the gnashing of teeth and constant
back and forth with Solidity over the number of variables in scope by
just deserializing directly into a struct and passing that around
directly to any function that needs initialization values, as opposed to
sending in the values separately. This requires encoding the struct as
bytes memory before sending it into the GovernorDiamondInit, but since
we can deserialize right back into a GovernorSettings struct, this is
trivial.